+++ sync-for-payables/src/codatsyncpayables/models/operations/delete_billpayment.py
@@ -8,24 +8,27 @@ from ..shared import pushoperation as shared_pushoperation
 from typing import Optional
 
 
-
 @dataclasses.dataclass
 class DeleteBillPaymentRequest:
     bill_payment_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'billPaymentId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a bill payment."""
     company_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'companyId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a company."""
     
-
 
 
 
 @dataclasses.dataclass
 class DeleteBillPaymentResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
     r"""Your API request was not properly authorized."""
     push_operation: Optional[shared_pushoperation.PushOperation] = dataclasses.field(default=None)
     r"""OK"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""