+++ sync-for-payables/src/codatsyncpayables/models/operations/get_journal.py
@@ -8,24 +8,27 @@ from ..shared import journal as shared_journal
 from typing import Optional
 
 
-
 @dataclasses.dataclass
 class GetJournalRequest:
     company_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'companyId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a company."""
     journal_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'journalId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a journal."""
     
-
 
 
 
 @dataclasses.dataclass
 class GetJournalResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
     r"""Your API request was not properly authorized."""
     journal: Optional[shared_journal.Journal] = dataclasses.field(default=None)
     r"""Success"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""